+++ x/evm/keeper/statedb.go
@@ -16,6 +16,7 @@
 package keeper
 
 import (
+	sdkstore "cosmossdk.io/store/types"
 	"fmt"
 	"math/big"
 
@@ -71,7 +72,7 @@ func (k *Keeper) ForEachStorage(ctx sdk.Context, addr common.Address, cb func(ke
 	store := ctx.KVStore(k.storeKey)
 	prefix := types.AddressStoragePrefix(addr)
 
-	iterator := sdk.KVStorePrefixIterator(store, prefix)
+	iterator := sdkstore.KVStorePrefixIterator(store, prefix)
 	defer iterator.Close()
 
 	for ; iterator.Valid(); iterator.Next() {